Move spawnable and states structs to an interface crate to later hoist certain plugins into separate crates. This aids in conditional compilation and platform-specific functionality. Wasm is a bear
