+++ Cargo.toml
@@ -3,6 +3,16 @@ name = "thetawave"
 version = "0.1.1"
 edition = "2021"
 
+[workspace]
+members = [
+    "crates/thetawave_interface"
+]
+
+[workspace.dependencies]
+serde = "1.0.159"
+strum = "0.25.0"
+strum_macros = "0.25.1"
+
 [dependencies]
 bevy = { version = "0.11.0", features = ["serialize"] }
 bevy_rapier2d = { version = "0.22.0", features = [
@@ -15,11 +25,11 @@ bevy_asset_loader = { version = "0.17.0", features = [
     "3d",
     "standard_dynamic_assets",
 ] }
-serde = "1.0.159"
+serde = {workspace = true}
+strum = {workspace = true}
+strum_macros = {workspace = true}
 ron = "0.8.0"
 rand = "0.8.5"
-strum = "0.25.0"
-strum_macros = "0.25.1"
 console_error_panic_hook = "0.1.7"
 bevy_kira_audio = { version = "0.16.0", features = ["mp3", "wav"] }
 #bevy_editor_pls = "0.4.0"
@@ -29,6 +39,7 @@ bytes = "1.4.0"
 rusqlite = { version = "0.29.0", features = ["bundled"] }
 dirs = "5.0.1"
 thiserror = "1.0"
+thetawave_interface = { path = "crates/thetawave_interface" }
 
 # Serial ports do not work in in the browser.
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
