+++ crates/thetawave_interface/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "thetawave_interface"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = {workspace = true}
+strum = {workspace = true}
+strum_macros = {workspace = true}
+bevy_ecs_macros = "0.11.0"
+bevy_ecs = "0.11.0"
