+++ src/collision/instersection.rs
@@ -1,12 +1,10 @@
 use crate::{
     player::PlayerComponent,
-    spawnable::{
-        ConsumableComponent, Faction, MobComponent, MobSegmentComponent, MobSegmentType, MobType,
-        ProjectileComponent, ProjectileType,
-    },
+    spawnable::{ConsumableComponent, MobComponent, MobSegmentComponent, ProjectileComponent},
 };
 use bevy::prelude::*;
 use bevy_rapier2d::{prelude::*, rapier::prelude::CollisionEventFlags};
+use thetawave_interface::spawnable::{Faction, MobSegmentType, MobType, ProjectileType};
 
 use super::{CollidingEntityPair, SortedCollisionEvent};
