+++ src/db/plugin.rs
@@ -1,7 +1,7 @@
 /// Exposes a single Plugin that links the game and our persistence layer.
 use crate::game::counters::current_game_metrics::{EnemiesKilledCounter, ShotCounters};
-use crate::states;
 use bevy::prelude::*;
+use thetawave_interface::states;
 
 use super::core::{get_db, setup_db, DEFAULT_USER_ID};
 use super::user_stats::{
