+++ src/spawnable/projectile/mod.rs
@@ -2,6 +2,7 @@ use bevy::prelude::*;
 use bevy_rapier2d::prelude::*;
 use serde::Deserialize;
 use std::{collections::HashMap, string::ToString};
+use thetawave_interface::spawnable::{ProjectileType, SpawnableType};
 
 use crate::{
     animation::{AnimationComponent, AnimationData},
@@ -9,7 +10,7 @@ use crate::{
     game::GameParametersResource,
     misc::Health,
     spawnable::InitialMotion,
-    spawnable::{ProjectileType, SpawnableBehavior, SpawnableComponent, SpawnableType},
+    spawnable::{SpawnableBehavior, SpawnableComponent},
     states::GameCleanup,
     HORIZONTAL_BARRIER_COL_GROUP_MEMBERSHIP, SPAWNABLE_COL_GROUP_MEMBERSHIP,
 };
